% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_power.R
\name{mdes}
\alias{mdes}
\title{Determine the minimum detectable effect size}
\usage{
mdes(desired_power = 0.8, num_df, den_df, alpha = 0.05)
}
\arguments{
\item{desired_power}{Defaults to .8, but this can be changed to test for the
minimum detectable effect size for other levels of power.}

\item{num_df}{Numerator degrees of freedom for the test.}

\item{den_df}{Denominator degrees of freedom for the test.}

\item{alpha}{Alpha level for the test.}
}
\value{
Returns an f-squared statistic representing the minimum detectable effect
size with the desired power, degrees of freedom, and alpha.
}
\description{
Given desired power, degrees of freedom, and alpha, this function returns the
minimum detectable effect size as an f-squared statistic.
}
