% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predictions.R
\name{model_predictions}
\alias{model_predictions}
\title{Get predictions from a linear model or generalized linear model}
\usage{
model_predictions(model, data = NULL, label = NULL, type = "response")
}
\arguments{
\item{model}{Name of the model that you'll use to generate predictions}

\item{data}{a dataframe containing cases for predictions. Must include all
regressors from model. Default is NULL with predictions returned for the current sample.}

\item{label}{A string label to append to variable names for predicted values,
CIs and SE. Default is NULL with no append}

\item{type}{'response' or 'link'. Used only for glm objects. see predict()}
}
\value{
Returns a dataframe of model predictions, standard errors, and
confidence intervals.
}
\description{
If no data are provided, modelPredictions returns a numeric vector predicted
values for the sample, functioning as a simple wrapper for fitted.values().
If a dataframe with new values for Xs are provided, modelPredictions adds
predicted values and SEs for these new data to the dataframe using predict()
from car package.
}
\details{
If Data=NULL, returns a numeric vector of predicted values for sample.
If Data are provided, adds four new columns at the front of the dataframe
These variables are named Predicted (prediced value), CILo (lower bound of
\itemize{
\item 1 SE from Predicted), CIHi (upper bound of + 1 SE), and SE (Standard error
of predicted value). NOTE: For GLM, +-1 SE are calculated on the link scale
and then converted to the response scale (which will be asymetric) if
type = response. If Label is not NULL, than Label is appended to end of these
four variable names.
}
}
